refactor(trials): add file extension on import paths
