+++ src/trials/AppendersTrial.ts
@@ -1,7 +1,7 @@
 import log4js from 'log4js';
 import { Client, Permissions, Guild, GuildChannel, ThreadChannel, CommandInteraction, ApplicationCommandData, ApplicationCommandPermissionData, CreateRoleOptions, Intents } from 'discord.js';
 import config from 'config';
-import { setContext } from '../discord/DiscordAppender';
+import { setContext } from '../discord/DiscordAppender.js';
 
 type GuildCommandInteraction = CommandInteraction & { guildId: string; }
 export interface DiscordBotConfig {
