+++ src/trials/index.ts
@@ -1,8 +1,8 @@
 import log4js from 'log4js';
 log4js.configure("config/log_mocha.json");
 
-//import * as trial from './WebServerTrial';
+//import * as trial from './WebServerTrial.js';
 //trial.webServerTrial();
 
-import { trial } from './DiscordTrial';
+import { trial } from './DiscordTrial.js';
 trial();
